+++ backend/fcm-server/src/main/java/news/bombom/notification/repository/ArticleArrivalNotificationFailedRepository.java
@@ -1,6 +1,6 @@
 package news.bombom.notification.repository;
 
-import news.bombom.notification.domain.ArticleArrivalNotificationFailed;
+import news.bombom.article.domain.ArticleArrivalNotificationFailed;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArticleArrivalNotificationFailedRepository extends JpaRepository<ArticleArrivalNotificationFailed, Long> {
