+++ backend/fcm-server/src/main/java/news/bombom/notification/repository/ArticleArrivalNotificationRepository.java
@@ -2,7 +2,7 @@ package news.bombom.notification.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import news.bombom.notification.domain.ArticleArrivalNotification;
+import news.bombom.article.domain.ArticleArrivalNotification;
 import news.bombom.notification.domain.NotificationStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,9 +11,10 @@ import org.springframework.data.repository.query.Param;
 public interface ArticleArrivalNotificationRepository extends JpaRepository<ArticleArrivalNotification, Long> {
 
     @Query("""
-            SELECT n 
-            FROM ArticleArrivalNotification n 
-            WHERE n.status IN :statuses AND (n.nextRetryAt IS NULL OR n.nextRetryAt <= :now)
-    """)
-    List<ArticleArrivalNotification> findRetryCandidates(@Param("statuses") List<NotificationStatus> statuses, @Param("now") LocalDateTime now);
+                    SELECT n 
+                    FROM ArticleArrivalNotification n 
+                    WHERE n.status IN :statuses AND (n.nextRetryAt IS NULL OR n.nextRetryAt <= :now)
+            """)
+    List<ArticleArrivalNotification> findRetryCandidates(@Param("statuses") List<NotificationStatus> statuses,
+                                                         @Param("now") LocalDateTime now);
 }
