+++ backend/fcm-server/src/main/java/news/bombom/notification/service/message/ChallengeTodoReminderMessageBuilder.java
@@ -1,7 +1,7 @@
 package news.bombom.notification.service.message;
 
 import java.util.Map;
-import news.bombom.notification.domain.ChallengeTodoReminderNotification;
+import news.bombom.challenge.domain.ChallengeTodoReminderNotification;
 import news.bombom.notification.domain.MemberFcmToken;
 import news.bombom.notification.domain.Notification;
 import news.bombom.notification.domain.NotificationPayloadType;
